[TG-19] test: [TG-19][TG-20] Modificat alguns tests i creat d'altres per a provar tota la funcionalitat.
Merge de feat/incidencies a feat/votes per a tenir els tests

+++ src/main/java/com/safesteps/backend/domain/incidents/projections/VoteCountDBProjection.java
@@ -4,5 +4,5 @@ public interface VoteCountDBProjection {
     Long getId();
     Long getPositiveVotes();
     Long getNegativeVotes();
-    Long getReliabilityIndex();
+    Double getReliabilityIndex();
 }
